requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.15.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-WTF==1.2.1
frozenlist==1.5.0
gotrue==2.9.3
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
packaging==24.1
pluggy==1.5.0
postgrest==0.17.2
propcache==0.2.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.0.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
storage3==0.8.2
supabase==2.9.1
supafunc==0.6.2
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
Werkzeug==3.0.4
WTForms==3.1.2
yarl==1.17.0
